+++ enpt/options/options_schema.py
@@ -53,6 +53,8 @@ enpt_schema_input = dict(
         type='dict', required=False,
         schema=dict(
             output_dir=dict(type='string', required=False),
+            output_format=dict(type='string', required=False, allowed=['GTiff', 'ENVI']),
+            output_interleave=dict(type='string', required=False, allowed=['band', 'line', 'pixel'])
         )),
 
     processors=dict(
@@ -130,6 +132,8 @@ parameter_mapping = dict(
 
     # output
     output_dir=('output', 'output_dir'),
+    output_format=('output', 'output_format'),
+    output_interleave=('output', 'output_interleave'),
 
     # processors > toa_ref
     path_earthSunDist=('processors', 'toa_ref', 'path_earthSunDist'),